requirements.txt
mlflow==2.20.0
numpy>=1.25.0
pandas>=1.5.3
scikit-learn>=1.3.0
python-dotenv>=1.0.0
torch>=2.2.0
torchvision>=0.15.0
torchaudio>=2.0.0